Change the app list with new design
Which includes just the app name with space and icon with state and
allocation numbers.

I decided that this should be a table based on the way headers and
rows are aligned with eachother. I have to modify some table styles
in cg-style to make it align with the new layout.