Make whole org quicklook clickable
For UI reasons, so you can click on the org and the row containing
the org separately.

I used a new class which will push out the div with negative margins
and padding. Had to include a method for clicking on just the `h2` for
the org name, as the link wasn't receiving the click.