Fix check style command for full git URL
The check style command is supposed to check that cloudgov-style is
installed directly from github rather then from npm or linked and
run it's build if so. This is so you can push up a version of the
dashboard with a custom cg-style version that hasn't been published
yet in order to allow for easily putting up testing branches and
having them build correctly.

This command wasn't working in that it was also running the cg-style
build on local branches because they had "github" in the local path(
usually because of go). This command specifies that cg-style has to
be installed from github.