Partial fix for mis-placed click even on overview
Before, if you clicked to the side of the org text, it would register
as an org click and go to the org page, which wasn't desired. Now
you have to click on the actual org text.

The bug where you open the org menu before you click still exists.