Get basic overview working
I created a new component for org overview, to match the way it's done
on the space and overview overviews. It's useful for the containers
to be the data holders that get the data from the stores.

I used the fetch all psaces for org function which was useful but has
to come after the base spaces are fetched, which is not great.