Show apps
I was able to successfully reuse space_quicklook as it will list out
the apps. This space quicklook shows only the app status, not more
details like memory required in teh org overview UI. I was thinking
of adding some property to space quicklook that allows the dev to
specify how each app should look, so in this case, would show a new
app quicklook item with a memory information.