Add new app quicklook for app lists
The app quick list is a new component that shows quick information
in a row/column fashion. It will not use a table, meaning the app list
will have to be modified a bit to accomodate it. It comes with different
configurations on what data is presented in it, currently it can show
just app state or app state and memory.