Fixed route and other error handling
A change, maybe in axios ajax module, maybe somewhere else changed
the error responses so the actual data resides in a `response` object
within the error object. This was breaking the error catching. The
fix is the pass the error `response` object to the `handleError`
function so the correct data is supplied to it so it can pass the
error correctly to the UI with the correct information.