Fix cf api tests with new error objects
The CF API should now use the new axios error format of having the
error data nested in the `response` object. Also changes the naming
convention of this fake error in an attempt to make it more clear what
it is in the tests. Only modified this where the tests were failing.

I also added a catch to promises in the cf api code, to ensure that
when a piece of code failed in the error or other handlers, it wouldn't
time out but would provide a good error message to the developer
testing.