Adds UI notificaitons for service form
To ensure user knows when the instance is being created and after it
was successfully created.

I decided on a simple timeout in the service instance store so it
only sets certain states after a certain amount of time. This ensures
minimal changes to how the components work. If the timeout was in
the jsx code, it would be complicated to figure out how to configure
the jsx code to wait to remove the instance form component.