Refactor the service instance panel on app page
I modified the confirmation UI to match the routes panel because
it should be consistent in design.

I had to add things to the compelx_list compoonent:
- titleElement: to help when you ahve a more complex title then just
a string such as a span with an href in it.
- emptyMessage: a custom message when the list is empty. I also added
functionality that things don't break with an empty list.

This seems to make complex list useful in all places.