Add unsubscribe to stores for tests
Create a clean Store for each test requires us to unregister from the
dispatcher. `unsubscribe()` can be called in test cleanup to achieve this.

There is a caveat. The Stores still create a singleton which is never
unregistered. This can cause some unexpected behavior in the tests where the
singleton store throws an error when an action is dispatched which seems
unrelated to your test setup. We should consider instantiating those singletons
at app startup instead of always creating them at require time.