Refactor usage and limits
Changes stat to new more constrained layout. Required some re-org
of thes stat component, like moving the title of it inside
the component to set it as a certain width.

Also fixed react error when ResourceUsage was passing a string to
secondaryInfo prop of Stat, which should be an Element.

Requires 18F/cg-style#ms-panel_refactor_usage_limits.