Show partial app while updating
Based on feedback from design, we want to show the value the user
updated as the app is officially updating.

I still think the ideal UI here would have some loading indicator
covering the value until it's officially updated on cloud foundry,
but this is the best option right now.