WIP Idea for how global error UI could work
Designed the data of how a global error UI could work. This uses a
more global system, where the errors are held in their own error
store.

As a developer, you decided in cf_api what data is important to the user
and if data failure shoudl show a generic error message. You can pass
a small string about what the problem is and what data is missing.

From there, it's dispatched as a special error event, stored in a list
of errors in the error store and displayed to the user from the main
container. I didn't design this yet, but there would be a dismiss
button to remove the errors or potentially a timeout. I would also
designthe error store to only display one error if there are over
a certain amount