Avoid two-way binding with value
Specifying a value for the FormElement should set the initial value, but changes
made by the user are meant to propagate upward via the `onValidate` handlers.
Updating UI state on every prop change would mean the data model would override
what the user is typing depending on parsing and validation rules which results
in a poor experience.