Move catches to the actions
And figure out all actions that need to happen for when a thing
fails to fetch. In the case of the app page, if the whole app doesn't
load, then there should be a connection error AND an error on the
actual app UI (otherwise it will say "no apps"). This is still figuring
out the pattern of global errors.