Click handler for refresh button
Supplies the Action component with what to do when the notification
refresh button is clicked. This is passed down from global error to
notification, as from notifications standpoint, the action could
be anything.

For now, global error is going to be dumb and just always supply a
refresh button, but it would be easy to configure something else based
on what the error is later.