Start testing server outside of test specs
This allows the test server to be executed with command line arguments. The
arguments will be spawned as a child process. This allows us to wrap commands
with a test server. The server is started, and connection information is
appended to the environment for the child process. Once the child process exits,
the testing server will be closed.

I call this a "sidecar" pattern based on Netflix's Prana[1] except that the
services provided by the sidecar is really only making the test server
available.

[1] http://techblog.netflix.com/2014/11/prana-sidecar-for-your-netflix-paas.html