Fixes infinite loading bug on services
The reason the bug was happening:
The service instance store had a `loading` state set to true even after
it was done loading. I couldn't reproduce, but this may have been the
case with the plan store too.

The stores were not manually emitting
a change on the received events. They assume the `merge` methods
would emit a change. But the `merge` functions only emit changes if
something actually changed. So in stores like the instance store, where
separate fetches are happening, I think the instances were already,
meaning nothing had to be merged, meaning emit change didn't get called.
this meant that the loading state was still true on the store.