Add user spaces endpoint
User spaces is the list of spaces for which the user has the SpaceDeveloper
role. We use this to build a roles object so we can quickly identify if the user
has the `space_developer` role in the space.

We don't include other roles for spaces or orgs, but any of these roles can be
fetched with similar API calls.

https://apidocs.cloudfoundry.org/253/users/list_all_spaces_for_the_user.html