Show alternative prompt in empty space
For the expressive empty state in an empty space, we give tips on how to create
a new app for your space. If the user is not a SpaceDeveloper in the space, these
tips don't apply because they don't have permission to create apps. We show
alternative tips in that situation.