Deprecate old roles function for new
Uses hasRole for everything, and ensure hasRole can take an array
of roles or a single role through overloading. Decided to do this
to keep things simple and not have multiple way to check roles.

Still have to add functionality to add org roles to the user just
as the space roles are.