Fix a few webpack problems and ensure prod
There's a problem with some 3rd-party modules where they use a
separate `main` field in package.json, such as `browserify`. This
was deprecated in new webpack but most modules haven't updated, so
it breaks them. The `mainFields` field in the webpack config adds
this back in some it's same functionality.

I added the `-p` to the prod build, as supposedly webpack does some
fancy things with this turned on like tree-shaking. This build will
uglify by default, so the plugin is no longer needed.