Try at tests
I want to make a request to the api server to change the variant so
the request data can change to not have apps. This would allow tests
to change conditions in the API server.

I can't get it to work. Doing an ajax request with axios from the
node context fails with ECONNREFUSED. Trying with fetch from inside
fails with "TypeError: Uncaught error: Cannot read property 'variant' of null"
that's returned from smocks running.

Is there a way to do this through the server API rather the ajax? How
would I even gain access to the server from here?