Cache static assets
Allow static assets to be cached by the browser, but ensure the browser
revalidates to make sure it has the latest javascript.

This should result in the browser requesting the asset with an `if-modified-`
header. The server will respond with 304 if the asset is still valid, or send
the asset if it's not.

It's important that the javascript is up to date. Otherwise it could mismatch
the server code and cause undefined behavior.