Remove dedent from functional tests for urls
Up until webdriverio 4.8.0, the use of dedent to construct urls did not
matter. However with this commit
https://github.com/webdriverio/webdriverio/commit/877e276fe4b5f03600a1d1d2fbeb3563398753c2
which changes how url works, it now keeps the literally newline
characters and will try to encode them into the url causing the url to
look like '/#%0A/org/orgguid'.

This problem only came up recently because 4.8.0 just came out and
according to our package.json we have `"webdriverio": "^4.6.1"`. This
means that upgrading to 4.8.0 on the next build was completely
acceptable.