Refactor mailer and templates
Now Mailer is simple. It just has one method where you sendMail and you
pass in your data.
We now test the SMTPMailer with "mailcatcher" by spinning up a docker
container. With mailcatcher I just test the cases of a valid send and an
invalid send (when the smtp server is down)

The templates are now in a single place and the mailer does not need to
know which templates it needs to know since we just pass the data to the
mailer now.
Now that the templates are in one place, we can test expected output of
the compiled templates and I have added the testdata files for the
expected compiled templates to test against.