Simplify rendering the saving component in users
So if a variable is undefined then React will not render it. So for
this code, we can just define saving `let saving;`. At that point,
`saving = undefined`. This means React will not render it. When we
want to show saving, we can set it to an element like so.

A couple of things you can't do in React:
- Set state in a component through direct access, such as
`this.state.something = something`. You have to use the `setState`
method, like `this.setState({something: something});`
- Set state in the render method. In react, `setState` will call
the render method of the component. So if you call `setState` in
render, you're essentially creating a never-ending loop of render.
In our code, state should only be set in the constructor and on the
change handler for when a store changes or when something is clicked.