Keep track of saving on user object
The merge function takes a key, and object and a function. If will
use the key to match the object with any existing object in the
store. We almost always use guid as they key because its the unique
identifier. So if you merge a user, and a user object with that guid
doesn't exist yet, it will add the user. If it does exist, it will
update that user with any changed data of the object you passed in.

The function is a callback function that will pass a true arg if
the merge function did change the object. This can be used to decide
when to `emitChange` (which will make the coomonent re-render).

This change the saving functionality to be on the user objects
individually rather then on the whole user store.