Do not allow role checkboxes to update in error cases
Solves: https://favro.com/card/1e11108a2da81e3bd7153a7a/18F-5842

The problem was that even though there was an error from trying to
associate/dissociate spaces roles or just dissociate org roles, the
checkboxes would still reflect because addedUserRoles / deletedUserRoles
actions would be called.

This PR moves addedUserRoles and deletedUserRoles to in another place in
the promise chain.
Now, after succeeding with deleteSpaceUserPermissions and
deleteOrgUserPermissions, the deletedUserRoles action is called in error
either, it will call errorRemoveUser (which will be used for raising
errors to UI)

The problem with the association of space roles was that it was not
passing the error back up like the org space association. This PR
mirrors that same behavior. (There is no errorAddUser action right now
so no need to call that.)

There are tests in this PR to test that in success cases addedUserRoles
and deletedUserRoles actions are called. Also the tests test to make
sure those actions are NOT called in the case of an error.