Disable HTTPS SSL check ONLY on local cf
In PCFDev, you need to use https endpoints for UAA. The problem is that
with PCFDev, the certs are local and invalid. So we need to disable
verifying them. Since this is a sensitive setting, this commit also adds
double checks for other security settings. If we are running against a
production CF env, we should not run without secure cookies for example.
If these checks fail, we fail to start the app.
Also, by adding an additional env var that focuses on the local env, we
don't have to modify the existing manifests which will enforce the
stricter security check.