Create servicesNetwork and servicesTestNetwork
servicesNetwork is used for runtime services needed by the app
In this case that's just redis

servicesTestNetwork is used for the dockertest instances
https://github.com/ory/dockertest

In addition, we mount the docker socket to the container so that the
dockertests can work when ran from the docker container. (allows the
container to create another container itself)

In addition, added a connectToNetwork function which only runs in the
case that we are creating containers inside containers. This mode is
only activated when the environment variable DOCKER_IN_DOCKER is set.
This function allows the container to use a hostname that it can easily
find.

This commit also uses the connectToNetwork logic for Redis

Also, this commit reorders the order from where it will pull the redis
address from. Now, it will look in REDIS_URI first, then the
VCAP_SERVICES, then just defaults to localhost. This is because with the
docker environment, the docker container has the VCAP_SERVICES already
loaded. But for the test, we need to set the uri via REDIS_URI. So we
have to reorder that.