Fix user state bug
Pairing with @el-mapache
we found that if you navigate to the org or space page from home, you
are able to successfully change people's roles. However, if you refresh
the page, it will not work because the currentUser context is not set
upon refresh.

This patch removes this bit which regulated state updates