Move ErrorMessage to SystemErrorMessage
The former ErrorMessage was specifically designed to say "The system
encountered an error" and also "try again". This is too specific for
ErrorMessage to be a generic component to render error messages.
Instead, the old ErrorMessage is moved to SystemErrorMessage and
ErrorMessage now simply renders the error message given to it.