app: add env panel with CRUD for user-defined env (#1243)
* app: add env panel with CRUD for user-defined env

Also move ErrorMessage to SystemErrorMessage.

The former ErrorMessage was specifically designed to say "The system
encountered an error" and also "try again". This is too specific for
ErrorMessage to be a generic component to render error messages.
Instead, the old ErrorMessage is moved to SystemErrorMessage and
ErrorMessage now simply renders the error message given to it.

* app: fix test where app error response not exists

* Move env-related components to env directory

* env: define default values to document state type

* env: move preformatted env JSON render to method

* env: move panel Header and Section to own files

Also move EnvPanel into env_panel folder (accessed via index.js).

Also rename env components to just JS files.

* env: remove use of deprecated styler

* env: move propTypes and defaultProps to top

* Move propTypes and defaultProps to top in err cmps

* Remove styler from err cmps

Also change displayStyle prop to just inline for ErrorMessage and set it
to false by default.

* Remove displayType=global from sys err message

It was unused across the application. Instead changed to an inline prop
which defaults to true in this component to keep BC with current uses.

* env: refactor action creators and store

Also:

- Render the <Panel /> directly in EnvPanel (because that's what it is -
  it's not an "EnvPanelContents").
- Show loading message when loading env.
- Show error message when env fails to load.
- Fix naming of action creator types. Note: types are prefixed with
  "ENV_" because when dispatched we ideally want a unique prefix.
- Add and export specific propTypes from app and env store so that
  components can be DRY and remove the chance that they depend on a
  slightly different prop type to another component that is using the
  same data. Note: this is a perfect example of where types (e.g. Flow
  or TS) come into play.