Make breadcrumbs component dumb
This also fixes the "calling setState on an unmounted component"
warning.

Also, instead of using CSS classes to target DOM nodes in integration
tests, we use `data-test` attributes. We can later strip these out of
the build using a babel plugin, if desired. Note: there are other ways
of doing this, but for the time being the `data-test` attribute best
matches the current code with the least number of changes and is clear
enough. It's also a recommended approach by Kent C Dodds, if that holds
any weight.